added font awesome icon and moved stars component to helpers
